csv변경 및 변경 사항을 영웅/몬스터의 스탯/스킬셋 데이터에 반영
- csv의 변경사항은 다음과 같음
  1. %를 없애고 대신 숫자로 표현 (ex: 15 => 0.15)
  2. 1열,2열,3열,4열 등 ~열을 나타내는 항목에 "열"문자를 없앰
  3. 명중률 등의 숫자들을 추후에 계산식에 오차가 없도록 100으로 나눠 표현
  4. 위 변경 사항에 따른 수정점에 CSVImporter 코드들이 잘 작동하도록 변경
  5. 위 변경 사항에 맞춰 영웅/몬스터의 스탯/스킬 능력치 코드 부분을 수정

+++ Assets/Scripts/HeroData.cs
@@ -22,16 +22,16 @@ public class HeroData : ScriptableObject
     public int Level => _level;
 
     // 회피
-    [SerializeField] private int _dodge;
-    public int Dodge => _dodge;
+    [SerializeField] private float _dodge;
+    public float Dodge => _dodge;
 
     // 방어력
     [SerializeField] private float _protection;
     public float Protection => _protection;
 
     // 명중률 보정
-    [SerializeField] private int _correction;
-    public int Correction => _correction;
+    [SerializeField] private float _correction;
+    public float Correction => _correction;
 
     // 치명타 확률
     [SerializeField] private float _critical;
@@ -74,7 +74,7 @@ public class HeroData : ScriptableObject
     [Header("사용 기술")]
     public HeroSkillData[] skills;
 
-    public void Initialize(string name, int hp, int dodge, float prot, int speed, int correct,
+    public void Initialize(string name, int hp, float dodge, float prot, int speed, float correct,
                            float crit, int minAtk, int maxAtk, float[] resists, HeroSkillData[] skillAssets)
     {
         _heroName = name;
